requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
dash==2.17.1
dash-bootstrap-components==1.6.0
plotly==5.22.0
gunicorn==22.0.0

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.36,<3.0
psycopg2-binary>=2.9.9

# ── Data processing ────────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
requests==2.32.3
